Break up the controller test
